Removed last bits of code pertaining to custom image importing
